Presentation/API/requirements.txt
###############################################
# FASTAPI / API
###############################################
fastapi==0.115.3
uvicorn[standard]==0.30.6
python-multipart==0.0.9
apscheduler==3.10.4


###############################################
# Pydantic
###############################################
pydantic==2.9.2
pydantic[email]==2.9.2

###############################################
# Banco de Dados
###############################################
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

###############################################
# Segurança / Autenticação
###############################################
bcrypt==4.2.1
PyJWT==2.9.0

###############################################
# LangChain (STACK 0.2.x – ALINHADA)
###############################################
langchain==0.2.12
langchain-community==0.2.11
langchain-ollama==0.1.3
langchain-chroma==0.1.4
langchain-huggingface==0.0.3


###############################################
# ChromaDB (Vector Store)
###############################################
chromadb==0.5.3

###############################################
# Embeddings
###############################################
sentence-transformers==3.0.1
torch==2.8

###############################################
# Utils
###############################################
python-dotenv==1.0.1
cachetools==5.4.0
loguru==0.7.2
requests==2.32.3
